@@ -31,58 +31,57 @@
   "private": true,
   "license": "UNLICENSED",
   "devDependencies": {
-    "@babel/core": "^7.16.12",
+    "@babel/core": "~7.17.10",
     "@babel/plugin-syntax-dynamic-import": "~7.8.3",
-    "@babel/plugin-transform-modules-commonjs": "^7.16.8",
-    "@babel/preset-env": "^7.16.11",
-    "@babel/preset-react": "^7.16.7",
-    "@babel/preset-typescript": "^7.16.7",
-    "@testing-library/jest-dom": "^5.16.1",
-    "@testing-library/react": "^12.1.2",
+    "@babel/plugin-transform-modules-commonjs": "~7.17.9",
+    "@babel/preset-env": "~7.17.10",
+    "@babel/preset-react": "~7.16.7",
+    "@babel/preset-typescript": "~7.16.7",
+    "@testing-library/jest-dom": "~5.16.4",
+    "@testing-library/react": "~13.2.0",
     "@trshcmpctr/eslint-config-jest": "workspace:*",
     "@trshcmpctr/eslint-config-react": "workspace:*",
     "@trshcmpctr/eslint-config-typescript": "workspace:*",
     "@trshcmpctr/eslint-config": "workspace:*",
     "@trshcmpctr/jest-stdout-reporter": "workspace:*",
     "@trshcmpctr/markdownlint-config": "workspace:*",
-    "@types/express": "~4.17.13",
-    "@types/jest": "^27.4.0",
-    "@types/node": "^16.11.21",
-    "@types/react-dom": "^17.0.11",
-    "@types/react": "^17.0.38",
-    "@typescript-eslint/eslint-plugin": "~5.13.0",
-    "@typescript-eslint/parser": "~5.13.0",
-    "babel-jest": "^27.4.6",
-    "babel-loader": "^8.2.3",
-    "babel-plugin-syntax-dynamic-import": "^6.18.0",
-    "css-loader": "^6.5.1",
-    "eslint-import-resolver-typescript": "~2.5.0",
+    "@types/jest": "~27.5.1",
+    "@types/node": "~16.11.35",
+    "@types/react-dom": "~17.0.17",
+    "@types/react": "~17.0.45",
+    "@typescript-eslint/eslint-plugin": "~5.23.0",
+    "@typescript-eslint/parser": "~5.23.0",
+    "babel-jest": "~28.1.0",
+    "babel-loader": "~8.2.5",
+    "babel-plugin-syntax-dynamic-import": "~6.18.0",
+    "css-loader": "~6.7.1",
+    "eslint-import-resolver-typescript": "~2.7.1",
     "eslint-plugin-eslint-comments": "~3.2.0",
-    "eslint-plugin-import": "~2.25.4",
-    "eslint-plugin-jest": "~26.1.1",
+    "eslint-plugin-import": "~2.26.0",
+    "eslint-plugin-jest": "~26.2.2",
     "eslint-plugin-jsx-a11y": "~6.5.1",
     "eslint-plugin-node": "~11.1.0",
-    "eslint-plugin-react-hooks": "~4.3.0",
-    "eslint-plugin-react": "^7.28.0",
-    "eslint-plugin-testing-library": "^5.0.4",
-    "eslint-webpack-plugin": "^3.1.1",
-    "eslint": "^8.7.0",
-    "html-webpack-plugin": "^5.5.0",
-    "jest-serializer-path": "^0.1.15",
-    "jest": "^27.4.7",
+    "eslint-plugin-react-hooks": "~4.5.0",
+    "eslint-plugin-react": "~7.29.4",
+    "eslint-plugin-testing-library": "~5.5.0",
+    "eslint-webpack-plugin": "~3.1.1",
+    "eslint": "~8.15.0",
+    "html-webpack-plugin": "~5.5.0",
+    "jest-environment-jsdom": "~28.1.0",
+    "jest-serializer-path": "~0.1.15",
+    "jest": "~28.1.0",
     "markdownlint-cli2": "~0.4.0",
     "rimraf": "~3.0.2",
-    "standard-version": "^9.3.2",
-    "style-loader": "^3.3.1",
-    "terser-webpack-plugin": "^5.3.0",
-    "typescript": "^4.5.5",
-    "webpack-cli": "^4.9.1",
-    "webpack-dev-server": "^4.7.3",
-    "webpack": "^5.67.0"
+    "style-loader": "~3.3.1",
+    "terser-webpack-plugin": "~5.3.1",
+    "typescript": "~4.6.4",
+    "webpack-cli": "~4.9.2",
+    "webpack-dev-server": "~4.9.0",
+    "webpack": "~5.72.1"
   },
   "dependencies": {
-    "core-js": "^3.20.3",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2"
+    "core-js": "~3.22.5",
+    "react-dom": "~18.1.0",
+    "react": "~18.1.0"
   }
 }
