@@ -26,27 +26,27 @@
     "test": "jest --no-cache --reporters=@trshcmpctr/jest-stdout-reporter --silent"
   },
   "devDependencies": {
-    "@trshcmpctr/eslint-config": "workspace:*",
     "@trshcmpctr/eslint-config-jest": "workspace:*",
+    "@trshcmpctr/eslint-config": "workspace:*",
     "@trshcmpctr/jest-stdout-reporter": "workspace:*",
     "@trshcmpctr/markdownlint-config": "workspace:*",
     "eslint-plugin-eslint-comments": "~3.2.0",
-    "eslint-plugin-jest": "~26.1.1",
-    "eslint-plugin-import": "~2.25.4",
+    "eslint-plugin-import": "~2.26.0",
+    "eslint-plugin-jest": "~26.2.2",
     "eslint-plugin-jsx-a11y": "~6.5.1",
     "eslint-plugin-node": "~11.1.0",
-    "eslint-plugin-react-hooks": "~4.3.0",
-    "eslint-plugin-react": "^7.28.0",
-    "eslint": "^8.7.0",
-    "jest": "^27.4.7",
+    "eslint-plugin-react-hooks": "~4.5.0",
+    "eslint-plugin-react": "~7.29.4",
+    "eslint": "~8.15.0",
+    "jest": "~28.1.0",
     "markdownlint-cli2": "~0.4.0",
     "rimraf": "~3.0.2",
-    "typescript": "^4.5.5"
+    "typescript": "~4.6.4"
   },
   "peerDependencies": {
-    "eslint": "^7.x.x || ^8.x.x",
-    "eslint-plugin-jsx-a11y": "^6.5.1",
-    "eslint-plugin-react": "^7.28.0",
-    "eslint-plugin-react-hooks": "^4.3.0"
+    "eslint-plugin-jsx-a11y": "~6.5.1",
+    "eslint-plugin-react-hooks": "~4.5.0",
+    "eslint-plugin-react": "~7.29.4",
+    "eslint": "~8.15.0"
   }
 }
