@@ -30,31 +30,31 @@
     "resolver": "<rootDir>/jestResolver.js"
   },
   "devDependencies": {
-    "@trshcmpctr/eslint-config": "workspace:*",
     "@trshcmpctr/eslint-config-jest": "workspace:*",
+    "@trshcmpctr/eslint-config": "workspace:*",
     "@trshcmpctr/jest-stdout-reporter": "workspace:*",
     "@trshcmpctr/markdownlint-config": "workspace:*",
-    "@typescript-eslint/eslint-plugin": "~5.13.0",
-    "@typescript-eslint/parser": "~5.13.0",
-    "enhanced-resolve": "~5.9.2",
-    "eslint-import-resolver-typescript": "~2.5.0",
+    "@typescript-eslint/eslint-plugin": "~5.23.0",
+    "@typescript-eslint/parser": "~5.23.0",
+    "enhanced-resolve": "~5.9.3",
+    "eslint-import-resolver-typescript": "~2.7.1",
     "eslint-plugin-eslint-comments": "~3.2.0",
-    "eslint-plugin-jest": "~26.1.1",
-    "eslint-plugin-import": "~2.25.4",
+    "eslint-plugin-import": "~2.26.0",
+    "eslint-plugin-jest": "~26.2.2",
     "eslint-plugin-node": "~11.1.0",
-    "eslint": "^8.7.0",
-    "jest": "^27.4.7",
+    "eslint": "~8.15.0",
+    "jest": "~28.1.0",
     "markdownlint-cli2": "~0.4.0",
     "rimraf": "~3.0.2",
-    "typescript": "^4.5.5"
+    "typescript": "~4.6.4"
   },
   "peerDependencies": {
-    "@typescript-eslint/parser": "~5.13.0",
-    "@typescript-eslint/eslint-plugin": "~5.13.0",
-    "eslint-import-resolver-typescript": "~2.5.0",
-    "eslint-plugin-import": "~2.25.4",
-    "eslint": "^7.x.x || ^8.x.x",
-    "eslint-plugin-jest": "~26.1.1",
-    "typescript": "^4.x.x"
+    "@typescript-eslint/eslint-plugin": "~5.23.0",
+    "@typescript-eslint/parser": "~5.23.0",
+    "eslint-import-resolver-typescript": "~2.7.1",
+    "eslint-plugin-import": "~2.26.0",
+    "eslint-plugin-jest": "~26.2.2",
+    "eslint": "~8.15.0",
+    "typescript": "~4.6.4"
   }
 }
