@@ -31,17 +31,17 @@
     "@trshcmpctr/jest-stdout-reporter": "workspace:*",
     "@trshcmpctr/markdownlint-config": "workspace:*",
     "eslint-plugin-eslint-comments": "~3.2.0",
-    "eslint-plugin-import": "~2.25.4",
-    "eslint-plugin-jest": "~26.1.1",
+    "eslint-plugin-import": "~2.26.0",
+    "eslint-plugin-jest": "~26.2.2",
     "eslint-plugin-node": "~11.1.0",
-    "eslint": "^8.7.0",
-    "jest": "^27.4.7",
-    "lint-staged": "~12.3.3",
+    "eslint": "~8.15.0",
+    "jest": "~28.1.0",
+    "lint-staged": "~12.4.1",
     "markdownlint-cli2": "~0.4.0",
     "rimraf": "~3.0.2",
-    "typescript": "^4.5.5"
+    "typescript": "~4.6.4"
   },
   "peerDependencies": {
-    "eslint": "^7.x.x || ^8.x.x"
+    "eslint": "~8.15.0"
   }
 }
