@@ -26,15 +26,15 @@
     "test": "true"
   },
   "dependencies": {
-    "@jest/reporters": "^27.4.6"
+    "@jest/reporters": "~28.1.0"
   },
   "devDependencies": {
     "@trshcmpctr/eslint-config": "workspace:*",
     "@trshcmpctr/markdownlint-config": "workspace:*",
     "eslint-plugin-eslint-comments": "~3.2.0",
-    "eslint-plugin-import": "~2.25.4",
+    "eslint-plugin-import": "~2.26.0",
     "eslint-plugin-node": "~11.1.0",
-    "eslint": "^8.7.0",
+    "eslint": "~8.15.0",
     "markdownlint-cli2": "~0.4.0",
     "rimraf": "~3.0.2"
   }
