@@ -31,14 +31,14 @@
     "@trshcmpctr/eslint-config": "workspace:*",
     "@trshcmpctr/markdownlint-config": "workspace:*",
     "eslint-plugin-eslint-comments": "~3.2.0",
-    "eslint-plugin-import": "~2.25.4",
+    "eslint-plugin-import": "~2.26.0",
     "eslint-plugin-node": "~11.1.0",
-    "eslint": "^8.7.0",
+    "eslint": "~8.15.0",
     "markdownlint-cli2": "~0.4.0",
-    "plop": "~3.0.5",
+    "plop": "~3.1.0",
     "rimraf": "~3.0.2"
   },
   "peerDependencies": {
-    "plop": "~3.0.5"
+    "plop": "~3.1.0"
   }
 }
