@@ -1,4 +1,6 @@
-const { getRenderLoginWithData } = require('.');
+import { jest } from '@jest/globals';
+
+import { getRenderLoginWithData } from './get-render-login-with-data';
 
 describe('getRenderLoginWithData creates a handler that', () => {
   const renderLogin = getRenderLoginWithData({ clientId: 'my-client-id' });
