@@ -1,4 +1,6 @@
-const { getFetchWithOauth } = require('.');
+import { jest } from '@jest/globals';
+
+import { getFetchWithOauth } from './get-fetch-with-oauth';
 
 /**
  * Fake implementation of fetch
@@ -13,13 +15,16 @@ describe('getFetchWithOauth', () => {
   let fetch;
   let fetchWithOauth;
 
-  beforeEach(() => {
-    fetch && fetch.mockClear();
+  beforeAll(() => {
     fetch = jest.fn(fetchSucceeds);
     fetchWithOauth = getFetchWithOauth(fetch, {
       token_type: 'Bearer',
       access_token: 'accesstoken'
     });
+  });
+
+  afterEach(() => {
+    fetch.mockClear();
   });
 
   it('calls fetch with url', async () => {
