@@ -8,4 +8,4 @@ const getFetchWithOauth = (fetch, { token_type: type, access_token: token }) => 
   return async url => await fetch(url, { headers: { authorization: `${type} ${token}` }});
 };
 
-module.exports = { getFetchWithOauth };
+export { getFetchWithOauth };
